requirements.txt
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-filter==25.2
drf-spectacular==0.29.0
channels==4.3.2 
channels-redis==4.3.0 
uvicorn[standard]
dj_database_url==3.1.0
pytz
whitenoise==6.11.0
psycopg2-binary==2.9.11
gunicorn
